v2

+++ move_utils.py
@@ -33,9 +33,9 @@ def tar_outputs(scan_date, dir_path, tag, outdir):
                 os.makedirs(os.path.join(cwd, scan_date, outdir))
 
             file_path = os.path.join(cwd, scan_date, outdir, f'{scan_date}_{d_type}_{tag}.tar') 
-
-            with tarfile.open(file_path, 'w') as tar:
-                tar.add(d_type, recursive=True)
+            if not os.path.isfile(file_path):
+                with tarfile.open(file_path, 'w') as tar:
+                    tar.add(d_type, recursive=True)
     os.chdir(cwd)
